Bug 1163182 - refactor bug suggestions artifact generation
